* master: (92 commits)
  update
  WP-348
  Fixed issue with log disabling + fixed issue with adding monolog handlers on the fly
  Cleaned up unused namespaces + fixed EntityAbstract::resultToEntity() method
  Fixed issues after review: removed setLogger() + refactored logger injecting into abstract classes + fixed "disable logging" option
  Fixed issues after review: removed setLogger() which is needed only for DI purpose
  Fixed issues after review
  MonologWrapper::getLogger() returns logger with channel = called class name
  Deleted logger.smartling_verbose_metadata_io option
  Deleted logger.smartling_log_raw_sql option
  Deleted logger.smartling_verbose_output_for_xml_coding option
  Fixed tests
  Logger in DownloadTranslationJob and LastModifiedCheckJob class
  Logger in SubmissionCollectorJob class
  Logger in UploadJob class
  Logger in DbMigrationManager class
  Logger in ContentSerializationHelper class
  Logger in FieldsFilterHelper class
  Logger in ContentHelper class
  Logger in TranslationHelper class
  ...

# Conflicts:
#	inc/Smartling/ApiWrapper.php
#	inc/Smartling/Base/SmartlingCore.php
#	inc/Smartling/Base/SmartlingCoreUploadTrait.php
#	inc/Smartling/Bootstrap.php
#	inc/Smartling/ContentTypes/CustomPostType.php

+++ inc/Smartling/Base/SmartlingCore.php
@@ -11,6 +11,7 @@ use Smartling\Exception\SmartlingExceptionAbstract;
 use Smartling\Helpers\CommonLogMessagesTrait;
 use Smartling\Helpers\EventParameters\ProcessRelatedContentParams;
 use Smartling\Helpers\StringHelper;
+use Smartling\MonologWrapper\MonologWrapper;
 use Smartling\Queue\Queue;
 use Smartling\Settings\ConfigurationProfileEntity;
 use Smartling\Submissions\SubmissionEntity;
@@ -27,12 +28,14 @@ class SmartlingCore extends SmartlingCoreAbstract
 
     public function __construct()
     {
+        parent::__construct();
+
         add_action(ExportedAPI::ACTION_SMARTLING_SEND_FILE_FOR_TRANSLATION, [$this, 'sendForTranslationBySubmission']);
         add_action(ExportedAPI::ACTION_SMARTLING_DOWNLOAD_TRANSLATION, [$this, 'downloadTranslationBySubmission',]);
         add_action(ExportedAPI::ACTION_SMARTLING_CLONE_CONTENT, [$this, 'cloneWithoutTranslation']);
         add_action(ExportedAPI::ACTION_SMARTLING_REGENERATE_THUMBNAILS, [$this, 'regenerateTargetThumbnailsBySubmission']);
         add_filter(ExportedAPI::FILTER_SMARTLING_PREPARE_TARGET_CONTENT, [$this, 'prepareTargetContent']);
-        add_filter(ExportedAPI::ACTION_SMARTLING_SYNC_MEDIA_ATTACHMENT, [$this, 'syncAttachment']);
+        add_action(ExportedAPI::ACTION_SMARTLING_SYNC_MEDIA_ATTACHMENT, [$this, 'syncAttachment']);
     }
 
     /**
@@ -127,18 +130,30 @@ class SmartlingCore extends SmartlingCoreAbstract
      */
     protected function sendFile(SubmissionEntity $submission, $xmlFileContent, array $smartlingLocaleList = [])
     {
-        $tmp_file = tempnam(sys_get_temp_dir(), '_smartling_temp_');
+        $workDir = sys_get_temp_dir();
 
-        file_put_contents($tmp_file, $xmlFileContent);
+        if (is_writable($workDir)) {
+            $tmp_file = tempnam($workDir, '_smartling_temp_');
+            $bytesWritten = file_put_contents($tmp_file, $xmlFileContent);
 
-        if (!StringHelper::isNullOrEmpty($submission->getJobId())) {
-            $smartlingLocaleList = [];
+            if (0 === $bytesWritten) {
+                $this->getLogger()->warning('Nothing was written to temporary file.');
+                return false;
+            } else {
+                $tmpFileSize = filesize($tmp_file);
+                if ($tmpFileSize !== $bytesWritten || $tmpFileSize !== strlen($xmlFileContent)) {
+                    $this->getLogger()->warning('Expected size of temporary file doesn\'t equals to real.');
+                    return false;
+                } else {
+                    $result = $this->getApiWrapper()->uploadContent($submission, '', $tmp_file, $smartlingLocaleList);
+                    unlink($tmp_file);
+                    return $result;
+                }
+            }
+        } else {
+            $this->getLogger()->warning(vsprintf('Working directory : \'%s\' doesn\'t seen to be writable',[$workDir]));
+            return false;
         }
-        $result = $this->getApiWrapper()->uploadContent($submission, '', $tmp_file, $smartlingLocaleList);
-
-        unlink($tmp_file);
-
-        return $result;
     }
 
     /**
